Struct heliotrope::SolrQuery

pub struct SolrQuery {
    // some fields omitted
}

Represents Solr query. You'll need to build the query and pass it to Solr to execute. This struct is immutable, ie returns modified clone of itself when building. This is done to enable chaining.

Methods

impl SolrQuery

fn new(query: &str) -> SolrQuery

Creates a new SolrQuery with only query term inside it

fn add_field(&self, field: &str) -> SolrQuery

fn to_pairs(&self) -> Vec<(String, String)>

Converts this query to a vector of pairs, suitable for URL percent encoding