Struct heliotrope::SolrQueryResponse

pub struct SolrQueryResponse {
    pub status: int,
    pub time: int,
    pub total: int,
    pub start: int,
    pub items: Vec<SolrDocument>,
}

Solr query response

Fields

status

HTTP status. When failed to connect, it will be 0 (zero).

time

Time it took to execute the request in milliseconds

total

Total number of rows found. Note that this will probably be different from returned subset of rows, because Solr will always use pagination

start

Rows offset (zero based)

items

Current page of found Solr documents