Crate heliotrope

Heliotrope is a Solr client for the Rust programming language.

It builds with Cargo. To use it in your project, add this to your Cargo.toml file:

[dependencies.heliotrope]
git = "https://github.com/Valve/heliotrope"

Indexing

Adding new document to solr

use url::Url;
use heliotrope::{Solr, SolrDocument, SolrString, SolrI64};

fn main(){
    let url = Url::parse("http://localhost:8983/solr/test/").unwrap();
    let solr = Solr::new(url);
    let mut document = SolrDocument::new();
    document.add_field("id", SolrI64(1));
    document.add_field("type", SolrString("Book".to_string()));
    document.add_field("title", SolrString("How to train your dragon".to_string()));
    document.add_field("body", SolrString("Vala Morgulis".to_string()));
    match solr.add(&document) {
        Ok(solr_response) => println!("{:?}", solr_response),
        Err(solr_error) => println!("{:?}", solr_error)
    }
    match solr.commit() {
        Ok(solr_response) => println!("{:?}", solr_response),
        Err(solr_error) => println!("{:?}", solr_error)
    }
}

Add and commit in one step

use url::{Url};
use heliotrope::{Solr, SolrDocument, SolrString, SolrI64};

fn main(){
    let url = Url::parse("http://localhost:8983/solr/test/").unwrap();
    let solr = Solr::new(url);
    let mut document = SolrDocument::new();
    document.add_field("id", SolrI64(2));
    document.add_field("type", SolrString("Book".to_string()));
    document.add_field("title", SolrString("The Great Gatsby".to_string()));
    document.add_field("body", SolrString("In my younger and more vulnerable years..".to_string()));
    match solr.add_and_commit(&document) {
        Ok(solr_response) => println!("{:?}", solr_response),
        Err(solr_error) => println!("Status: {}, Message: {}", solr_error.status, solr_error.message)
    }
}

Adding multiple document at once

use url::{Url};
use heliotrope::{Solr, SolrDocument, SolrString, SolrI64};

let url = Url::parse("http://localhost:8983/solr/test/").unwrap();
let solr = Solr::new(url);
let mut document1 = SolrDocument::new();
document1.add_field("id", SolrI64(3));
document1.add_field("type", SolrString("Book".to_string()));
document1.add_field("title", SolrString("The Great Gatsby".to_string()));
document1.add_field("body", SolrString("In my younger and more vulnerable years".to_string()));

let mut document2 = SolrDocument::new();
document1.add_field("id", SolrI64(4));
document1.add_field("type", SolrString("Book".to_string()));
document2.add_field("title", SolrString("Moby Dick".to_string()));
document2.add_field("body", SolrString("Call me Ishmael".to_string()));

match solr.add_many_and_commit(vec!(&document1, &document2)) {
    Ok(solr_response) => println!("{:?}", solr_response),
    Err(solr_error) => println!("Status: {}, Message: {}", solr_error.status, solr_error.message)
}

Querying

extern crate url;
use url::{Url};
use heliotrope::{Solr, SolrDocument, SolrQuery, Descending};

let url = Url::parse("http://localhost:8983/solr/test/").unwrap();
let solr = Solr::new(url);
let query = SolrQuery::new("*:*").add_sort("year", Descending);
match solr.query(&query) {
    Ok(solr_response) => {
        println!("Status: {}", solr_response.status);
        println!("Time: {}", solr_response.time);
        println!("Total rows found: {}", solr_response.total);
        println!("Offset: {}", solr_response.start);
        for item in solr_response.items.iter() {
            println!("{}", item);
        }
    }
    Err(solr_error) => println!("Status: {}, Message: {}", solr_error.status, solr_error.message)
}

You can chain query options:

let query = SolrQuery::new("*:*")
    .add_field("score")
    .add_field("*")
    .add_sort("age", Descending)
    .add_sort("balance", Ascending);

Reexports

pub use document::{SolrF64, SolrI64, SolrString};

pub use query::{Ascending, Descending};

Structs

Solr
Represents your API connection to Solr. You will use this struct to perform operations on Solr.

SolrDocument
SolrDocument to be used to either index or query.

SolrError
SolrError

SolrField
SolrDocument field

SolrQuery
Represents Solr query. You'll need to build the query and pass it to Solr to execute. This struct is immutable, ie returns modified clone of itself when building. This is done to enable chaining.

SolrQueryResponse
Solr query response

SolrUpdateResponse
Solr response used for update/indexing/commit operations

Enums

SolrValue
Strongly typed Solr document field value

Type Definitions

SolrQueryResult

SolrUpdateResult